notifyicon.go: add support for successfully creating NotifyIcons when the taskbar is not present.
Until now, when the taskbar is not present (perhaps the application has been
started before the shell), creating a `NotifyIcon` will fail.

We can detect the presence of the taskbar by calling `SHAppBarMessage`.
If we determine that the taskbar is not present, we still successfully
create the `NotifyIcon` and the application may continue to mutate the
icon's state.

Since we already handle `TaskbarCreated` messages, we leverage that
existing functionality to then populate the notification icon once
the shell has started.

One problem that this creates is that, until now, `NotifyIcon` always
assumed that it had a valid notify icon ID. I refactored its
implementation to make it compatible with the notion that it may not
yet have an ID.

I added additional separation between the framework state tracked by
`NotifyIcon` and the state+execution of the shell APIs themselves:

A new struct, `shellNotificationIcon`, hosts the ID and HWND for the
icon. To mutate state, we call `shellNotificationIcon.newCmd` to create
a `niCmd` struct that executes mutations.

If the notification icon does not yet exist, `newCmd` returns a `nil`
`niCmd`. However, this is non-fatal and the `NotifyIcon` continues to
hold updated application state.

Notice that the `NOTIFYICONDATA.CbSize` field is now set to the
full size of that structure: I decided that I wanted to use the
HBalloonIcon field for balloon icons. This is still compatible with
the Windows XP compatible `NIM_SETVERSION` stuff: It is quite reasonable
to request Windows XP semantics for the notification icon's messaging
while simultaneously using a newer revision of the `NOTIFYICONDATA`
structure. The messaging semantics for the notification icon are
orthogonal to the ABI revision of the `NOTIFYICONDATA` structure.

As an added bonus, this means that we only need one call to
`Shell_NotifyIcon` to create a balloon notification with a user-defined icon.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>