notifyicon.go: change Shell_NotifyIcon to use win.lastError instead of win.newError
MSDN docs really suck here, as they do not indicate that `GetLastError` is valid
when `Shell_NotifyIcon` returns false. However, the disassembly indicates that
it is in fact used.

Let's update this so that we can get better errors when the call fails.

Updates https://github.com/tailscale/tailscale/issues/4133

Signed-off-by: Aaron Klotz <aaron@tailscale.com>