action.go: guard against freeing of reserved Action IDs
Separators were being created with a zero ID, which wasn't a big deal when we
weren't reusing IDs, but now that we are, we don't want that (since zero is
actually reserved). I added a guard to the allocator to prevent freeing of
reserved IDs.

Updates https://github.com/tailscale/corp/issues/8716

Signed-off-by: Aaron Klotz <aaron@tailscale.com>