dpicache: add dpicache package
The dpicache package implements a mapping from a canonical value to equivalent
values at specified DPIs. The DPI-specific values are lazily created upon
request by calling a CopyForDPI method on the canonical value.

For example, a key could consist of a structure containing UI metrics at 96DPI.
Calling InstanceForDPI with that key will return a copy of that structure with
metrics that are scaled to the requested DPI.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>