bufferedpaint.go: add support for buffered paint API
The buffered paint API takes care of management of back buffers, blitting from
the back buffer to the front buffer, and ensuring that the buffer is a top-down
DIB, which is needed for alpha blending.

It adds an invocation of BufferedPaintInit on the UI thread to set up a buffer
pool that is thread-local.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>