flesh out enough GDI+ support for owner-drawn menus
In order to properly support graphics with alpha blending and antialiasing, we
need something better than plain old GDI. Given the current state of COM-like
bindings, Direct2D is not (yet) an option. GDI+ is a reasonable short-term
compromise.

Note that this implementation does not replace Canvas's implementation,
as that would likely break compatibility with existing code in a visual sense.
Instead, a GDIPlusCanvas may be obtained from a Canvas and operations may then
be performed from there.

I did not implement the entire GDI+ API surface, but rather enough to meet
Tailscale's needs for painting menus.

The documentation does not delve deeply into nuances of GDI+ semantics, as it
seems redundant to me to rewrite Microsoft documentation. Instead, when the
semantics of a method are not obvious, I link to the relevant Microsoft docs.

Some parts of walk's Bitmap implementation already used GDI+. Those parts have
been refactored to use the new API bindings.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>