notifyicon.go: post WM_NULL after TrackPopupMenuEx returns
This is a known issue that, save for a few old-school Windows hackers, has been
lost to the dustbin of history.

When calling TrackPopupMenu(Ex) for notification icons, we need to do a little
dance to ensure that focus arrives and leaves the context menu correctly. The
original source for this information is long gone, but fortunately I found an
archived copy.

See https://web.archive.org/web/20000205130053/http://support.microsoft.com/support/kb/articles/q135/7/88.asp

Signed-off-by: Aaron Klotz <aaron@tailscale.com>