menu.go, window.go: always call win.DrawMenuBar, but only do so once for menus that are not the main menu
Apparently some RDP connections are not drawing context menus without HMENU
mutations being followed by an explicit call to win.DrawMenuBar even though
context menus are not part of a window's menu bar.

Originally walk was redrawing the menu bar after every single modification to an
HMENU, which was a performance disaster that scaled linearly with the number
of menu items in the menu. (Add 200 machines to the device menu? Walk would
repaint the menu bar 200 times.) In a previous PR I changed this to only
redraw the menu bar for modifications to a menu when that menu is acutally part
of the menu bar itself. That change worked great locally, but apparently it
messed up menu painting over RDP.

This edition on the fix sets a flag on a WindowBase indicating that a menu
redraw is necessary. We check that flag and perform the redraw as the menu is
in the process of being displayed, thus ensuring only a single DrawMenuBar call
that will capture all the previous mutations.

I also fixed a location where we should have been triggering a repaint but
weren't.

I also fixed a regression: (*Menu).onInitPopup should be called with
WindowBase.window as its argument, not the *WindowBase itself. The latter might
cause issues for types implementing Window that have their own implementation
of Window methods.

Updates https://github.com/tailscale/tailscale/issues/7698
Fixes #30

Signed-off-by: Aaron Klotz <aaron@tailscale.com>