form.go, notifyicon.go, window.go: improve robustness of (*NotifyIcon).reAddToTaskbar
Walk watches for a notification from Windows when the taskbar is created. The
idea is if Explorer crashes or the walk app starts ahead of Explorer, then
walk will know when it should try re-adding notification icons to the taskbar.

Unfortunately the taskbar also exhibits this behaviour during a change in
the primary display's pixel density. This can happen when an RDP connection is
made, or when the device's monitor configuration changes. In this case,
the Explorer process never terminated, so when we try to re-add our icons,
the Shell_NotifyIcon call fails because the icons are technically still there!

We deal with this nonsense by simply failing out if the Shell_NotifyIcon call
fails; the icon presumably hasn't changed. OTOH, if the call succeeds, we
check the ID against the previous ID, and if there is a change, we update our
understanding of ID mappings accordingly.

Local testing confirmed this fix.

Updates https://github.com/tailscale/tailscale/issues/7698

Signed-off-by: Aaron Klotz <aaron@tailscale.com>