event.go, proceedevent.go, taskdialog.go, examples/taskdialog: add TaskDialog support
Task Dialogs are the Vista-era successor to MessageBox: a simplified API for
displaying pre-defined dialog boxes.

This patch supports almost all functionality offered by the underlying API,
with the exception of "navigation," which produces wizard-like sequencing
of forms within the same dialog box. That might be added in the future,
depending on our needs.

The example program has been written to exercise much of the interface
implemented by this patch; note that it includes a manifest, which is
required in order to reference the correct version of comctl32.dll that we need
for TaskDialog support.

I also added variants of Event and ProceedEvent that support a generic argument
to simplify event handling code and facilitate reuse.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>