mainloop_default.go, mainloop_cgo.go, windowgroup.go, window.go: use per-WindowGroup, hidden, message-only window to process synchronization messages
Directly calling (*WindowGroup).RunSynchronized from the message pump will not
always work as expected: consider a nested message loop caused by something
like a displaying menu or a dialog box. Sync messages are sent to the walk
window and are dispatched to it, however that triggers an iteration of the
*nested* message loop, not the outer *walk* message loop!

The way to fix this is to create a hidden, message-only window that can
actually handle the sync message and dispatch its funcs.

Fixes https://github.com/tailscale/walk/issues/45

Signed-off-by: Aaron Klotz <aaron@tailscale.com>