menu.go: do some cleanup and add size invalidation to owner-drawn menu items
Owner-drawn menu items aren't automagically resized if necessary when their
content changes. Short of removing and then re-inserting the menu item, it
looks like the only way to invalidate the item's size is to specify the
MIIM_BITMAP flag with a zero bitmap handle. Don't ask me how I learned that.

Furthermore, we need to be careful about when we invalidate the item's size;
we pretty much only want to do it if the menu is about to be shown but has
not yet been displayed; if it happens during displaying, the menu will resize
itself, which is very disorienting visually.

I also did a bit of cleanup: eg the Cch field is not used for string items.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>