notifyicon.go: change behaviour of left-clicks when no mouse-up event handlers are registered
Windows 7 UX guidelines for notification icons state that, when there is no
other UI to be displayed upon left-click, then the default action should be
to display the context menu. We use the presence of event handlers as a
proxy to determine whether to perform the fallback.

Updates https://github.com/tailscale/corp/issues/14362

Signed-off-by: Aaron Klotz <aaron@tailscale.com>