action.go, menu.go: remove the MFT_OWNERDRAW flag when an owner-drawn menu item is converted to a standard menu item by Action.SetOwnerDraw(nil)
We need to raise the action changed event when ownerDrawInfo is being changed in Action.SetOwnerDraw, and have the menu.onActionChanged handler initialize and set MENUITEMINFO for the item. Additionally, it is important that the MIIM_FTYPE and MIIM_DATA flags are always set in initMenuItemInfoFromAction. There are no performance implications, and it allows us to properly unset the MFT_OWNERDRAW and MFT_RADIOCHECK flags when a menu item is no longer owner-drawn or no longer a radio button, respectively. This also enables us to clear the DwItemData field when it is no longer used.

Updates tailscale/corp#16659

Signed-off-by: Nick Khyl <nickk@tailscale.com>