window.go: Refer directly to appSingleton when resolving Application during window class registration
walk.App() is sometimes poorly suited for internal references to the app
because it includes a bunch of checks to ensure that it is not being called
too early in the application's lifecycle. On the other hand, internally
to the framework, we sometimes need to be able to reference the app without
those additional checks. This is one of those cases.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>