application.go, window.go: unconditionally return singleton pointer from App
During the review of a previous PR, we discussed altering the semantics
of App() such that it returns nil if called after InitApp() has failed.
Unfortunately Walk's pre-existing interdependencies are making that very
difficult to achieve in the sort term. I am reverting that change with
the intention of revisiting this in issue 61.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>