fix up some post-refactoring issues
I fixed a couple of things that were bothering me in action.go and
window.go, but the main thing that is actually a more significant bug is
that FormBase's OnPostDispatch was not being invoked as frequently as it
was pre-refactor.

There are two reasons for this. The first is that we should call the
handler when IsDialogMessage returns true, as that should be treated as
as dispatch. Secondly, we also need to handle messages that were dispatched
to child windows of the form. In the case of issue 70, the toolbar was
receiving a WM_COMMAND whose handler affected layout of the entire form.
In that case, we implement OnPostDispatch on WidgetBase and forward it
to the Form at the top of the hierarchy.

Fixes #70

Signed-off-by: Aaron Klotz <aaron@tailscale.com>