notifyicon.go: prevent attempts to reinitialize and show a context menu while it's already being displayed
We should not publish the
(*NotifyIcon).ShowingContextMenu event if the menu
is already being displayed. Otherwise, event
handlers might reinitialize the menu and update
the items, invalidating internal structures like
ownerDrawnMenuItemInfo, which are already in use.
This can cause various issues, including a nil
pointer dereference in (*menuItemLayout).layout().

Under normal circumstances, clicking on the tray
icon while a context menu is displayed should
close the context menu and make TrackPopupMenuEx
return zero before we receive WM_CONTEXTMENU or
WM_LBUTTONUP again. However, this doesn't happen in
certain scenarios, such as when clicking on the
tray icon immediately after opening and closing
the Start Menu. In these cases, SetForegroundWindow
fails, the context menu never receives focus, and
it appears in the background. Clicking outside of
the context menu (e.g. on the tray icon or on the
task bar) does not cancel the context menu. This
behavior is not unique to the Tailscale tray icon
and appears to be a Windows bug. While we cannot
fix the underlying issue that prevents tray menus
from becoming foreground and active, we can
prevent the ShowingContextMenu event from being
published and the walk structures from being
invalidated while the context menu is displayed,
fixing the nil dereference and similar issues.

In this PR, we add a field to track if we're in
(*NotifyIcon).doContextMenu and check it to
prevent doContextMenu from being called
recursively. We also update (*NotifyIcon) wndProc
to forcibly cancel and hide the context menu if a user
clicks on the tray icon while the context menu is
displayed.

Fixes #72

Signed-off-by: Nick Khyl <nickk@tailscale.com>