add support for immediate layout and plumb completion function throughout layout
Sometimes the application programmer wants layout to occur sooner than
walk wants to do it. Furthermore, the app might want to know when layout
has completed.

In this PR we add support for immediate layout, and we also add new APIs
that allow us to specify a callback function to notify us when layout
is complete.

Fixes https://github.com/tailscale/walk/issues/83

Signed-off-by: Aaron Klotz <aaron@tailscale.com>