application.go, minwin.go, win32window.go: change how walk surfaces panics, introduce MinWin
I've decided that I think walk would be better off surfacing panics
immediately, instead of giving users the chance to ignore/suppress them.
To make this happen in a way that won't entangle the panic with other,
possibly native code on the call stack, I use a new goroutine to execute
the panic.

Secondly, I wrote a simple HWND wrapper called MinWin (tongue planted
firmly in cheek), whose purpose is to provide the absolute minimum
windowing functionality needed for hosting XAML islands. Because the
MinWin is partially responsible for non-client stuff, I enhanced
Win32Window to support setting and retrieval of DWM attributes. I'm not
sure whether we'll need these or not, but I wanted to get this
functionality in ahead of time.

Fixes #101

Signed-off-by: Aaron Klotz <aaron@tailscale.com>