Make WindowBase aware of high-contrast mode and use COLOR_WINDOW as its background when enabled
We create a singleton for the COLOR_WINDOW system brush.

We update (*WindowBase).backgroundEffective() to always return said brush
when high-contrast is enabled. We also add some additional documentation
to that method to help explain its purpose.

Note that WM_SETTINGCHANGE always invalidates windows
(via ApplySysColorser), so we should expect this to update on the fly
on the rare occasion when walk windows are visible and high contrast mode
is toggled.

Updates https://github.com/tailscale/corp/issues/23789

Signed-off-by: Aaron Klotz <aaron@tailscale.com>