application.go: Add Go method and WaitGroup to Application
Sometimes when quitting a walk application built in debug mode, I see
panics from goroutines that have not finished running by the time the
process terminates. While sometimes this doesn't matter, there are
other times when we'd like to ensure that goroutines have finished.

I added a new Go method to Application that spawns goroutines that
will be associated with a WaitGroup and the Application's Context.
(*Application).Run is modified to wait on that WaitGroup before returning.

I also updated layout code to use (*Application).Go instead of
directly spawning goroutines.

Fixes #110

Signed-off-by: Aaron Klotz <aaron@tailscale.com>