form.go, notifyicon.go: Improved tracking of modal loops
For Form, we add a new event that is published if EnterMode is called.

For NotifyIcons, we move tracking whether existing context menus are
active to handlers for WM_ENTERMENULOOP and WM_EXITMENULOOP, as these
messages give us more precise notification as to when menu(s) are
present. We also change the tracking variable from a bool to an int,
since Win32 APIs to permit nested context menus when specific flags
are provided.

We also pick up the latest revision of win.

Updates https://github.com/tailscale/corp/issues/26617

Signed-off-by: Aaron Klotz <aaron@tailscale.com>