notifyicon.go: a11y improvements to notification icons
We handle NIN_KEYSELECT and NIN_SELECT identically to WM_LBUTTONUP.

We initialize the hidden event sink window with WS_DISABLED and annotate
it with a window role instead of a client role, in the hope that this
will discourage screen readers from focusing on it.

Finally, we do a bit of cleanup since we're already in here.

Updates https://github.com/tailscale/corp/issues/29972

Signed-off-by: Aaron Klotz <aaron@tailscale.com>